// frontend/src/shared/api/notes/model/note.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Notes & Recap API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteScope } from './noteScope';

export interface Note {
  id: number;
  show_id: number;
  scope: NoteScope;
  /** @nullable */
  season_number?: number | null;
  /** @nullable */
  episode_number?: number | null;
  body: string;
  created_at: string;
}